% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pridit.R
\name{coef.pridit}
\alias{coef.pridit}
\title{Extract PRIDIT weights}
\usage{
\method{coef}{pridit}(object, ...)
}
\arguments{
\item{object}{A \code{"pridit"} object.}

\item{...}{Ignored.}
}
\value{
Named numeric vector of PRIDIT weights.
}
\description{
Extract PRIDIT weights
}
