% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIDITPackage.R
\docType{package}
\name{pridit}
\alias{pridit-package}
\alias{pridit}
\title{pridit: Principal Component Analysis Applied to Ridit Scoring}
\description{
The pridit package provides functions for implementing the PRIDIT
(Principal Component Analysis applied to RIDITs) scoring system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rlieberthal/PRIDIT}
  \item Report bugs at \url{https://github.com/rlieberthal/PRIDIT/issues}
}

}
\author{
\strong{Maintainer}: Robert D. Lieberthal \email{rlieberthal@gmail.com}

}
