% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIDITscore.R
\name{PRIDITscore}
\alias{PRIDITscore}
\title{Calculate the PRIDIT scores for a ridit matrix}
\usage{
PRIDITscore(riditscores, id_vector, weightvec)
}
\arguments{
\item{riditscores}{A matrix where the first column represents IDs.
The IDs uniquely identify each row in the matrix.
The remaining columns contain the ridit scores for each ID. # Inline comment}

\item{id_vector}{A vector of IDs.}

\item{weightvec}{A vector of PRIDIT weights.}
}
\value{
A data frame with the following columns:
  \describe{
    \item{ID}{The unique identifier for each row.}
    \item{ScoreMatrix}{A matrix containing PRIDIT scores for each 
    observation.}
  }
}
\description{
This function takes a matrix of data and returns the matrix transformed 
as ridit values.
}
