% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIDITweight.R
\name{PRIDITweight}
\alias{PRIDITweight}
\title{Calculate the PRIDIT weights for a ridit matrix}
\usage{
PRIDITweight(riditscores)
}
\arguments{
\item{riditscores}{A matrix where the first column represents IDs.
The IDs uniquely identify each row in the matrix.
The remaining columns contain the data for each ID.}
}
\value{
A data frame with the following columns:
  \describe{
    \item{ID}{The unique identifier for each row.}
    \item{WeightMatrix}{A matrix containing PRIDIT weights for each variable.}
  }
ridit(2)
}
\description{
This function takes a matrix of data and returns PRIDIT weights for
each variable as a vector.
}
