% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridit.R
\name{ridit}
\alias{ridit}
\title{Calculate the ridit values for a matrix}
\usage{
ridit(allrawdata)
}
\arguments{
\item{allrawdata}{A matrix where the first column represents IDs.
The IDs uniquely identify each row in the matrix.
The remaining columns contain the data for each ID.}
}
\value{
A data frame with the following columns:
  \describe{
    \item{ID}{The unique identifier for each row.}
    \item{DataMatrix}{A matrix containing additional data columns.}
  }
ridit(2)
}
\description{
This function takes a matrix of data and returns the matrix transformed 
as ridit values.
}
